Move non-task functions into their own module
